refactor(app): reorganize contribution editor structure
Decompose the contribution markdown editor into smaller, modular components. Custom toolbar popovers and buttons are extracted into EditorToolbar, and the range-based shortcut listener is extracted into MarkdownLinkImageShortcutListener. The editor enhancements markdown document is moved to the centralized docs folder.
